Move the current user logic to its own handler separate from the endpoint handlers
That way each individual endpoint handler doesn't need to also worry about user binding logic - it can now be done once at the app level and the endpoint handlers can be more pure as a result.
